Revert "Merge github.com:torvalds/linux into 5.16"
This reverts commit 584d795d1ee7c9b1f5b14ba4e254cf1161ae2ee6, reversing
changes made to ec182db76a9a39dfd6b7119c4cdb0d46c086dc7f.

This could be the problem as well

+++ fs/xfs/xfs_ioctl.c
@@ -372,7 +372,7 @@ int
 xfs_ioc_attr_list(
 	struct xfs_inode		*dp,
 	void __user			*ubuf,
-	size_t				bufsize,
+	int				bufsize,
 	int				flags,
 	struct xfs_attrlist_cursor __user *ucursor)
 {
